refactor: перенести файл reverse.ts и sorting-page.tsx, подготовить файлы к удвлению пакета uuid// move the files reverse.ts and sorting-page.tsx, prepare the files for removing the uuid package
