+++ src/components/app/app.tsx
@@ -5,7 +5,7 @@ import { ListPage } from "../list-page/list-page";
 import { MainPage } from "../main-page/main-page";
 import { QueuePage } from "../queue-page/queue-page";
 import { StringComponent } from "../../pages/string/string";
-import { SortingPage } from "../sorting-page/sorting-page";
+import { SortingPage } from "../../pages/sorting-page/sorting-page";
 import { StackPage } from "../stack-page/stack-page";
 
 import "./app.css";
