--- src/components/sorting-page/sorting-page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { SolutionLayout } from "../ui/solution-layout/solution-layout";
-
-export const SortingPage: React.FC = () => {
-  return (
-    <SolutionLayout title="Сортировка массива">
-
-    </SolutionLayout>
-  );
-};
